Rename 'mount' module to 'decorators'
